requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
httpx>=0.25.2,<0.26.0
sse-starlette==1.6.5
aiohttp==3.9.1
wbgapi==1.0.3
pandas==2.2.3
mistralai==0.1.2
anyio>=4.5.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1